+++ lib/libexec/grammar.sh
@@ -55,6 +55,7 @@ mapping() {
     Before      )   control before      "$2" ;;
     After       )   control after       "$2" ;;
     Pending     )   control pending     "$2" ;;
+    Set         )   control set         "$2" ;;
     Skip        )   skip                "$2" ;;
     Intercept   )   control intercept   "$2" ;;
     Data        )   data raw            "$2" ;;